Move contents of utils to lib
