+++ src/components/Carousel/Stage/index.spec.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "@jest/globals";
 import { screen } from "@testing-library/dom";
 
-import createTestRender from "@/utils/createTestRender";
+import createTestRender from "@/lib/createTestRender";
 
 import { imageA, imageB, imageC } from "../test-data";
 import CarouselStage from ".";
